refactor: put `http_client` traits in dedicated directory
